\name{xsystem-package}
\alias{xsystem-package}
\alias{xsystem}
\docType{package}
\title{
\packageTitle{xsystem}
}
\description{
\packageDescription{xsystem}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{xsystem}
\packageIndices{xsystem}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{xsystem}

Maintainer: \packageMaintainer{xsystem}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
